Updated tests file structure.
